Complete web-ui consolidation - move all files to root level
