+++ package-backup.json
@@ -1,0 +1,17 @@
+{
+  "name": "visio-temp-file-remover",
+  "version": "1.0.0",
+  "description": "Utility to find and delete temporary Visio files",
+  "main": "app.js",
+  "scripts": {
+    "start": "node app.js",
+    "dev": "nodemon app.js"
+  },
+  "dependencies": {
+    "express": "^4.18.2",
+    "body-parser": "^1.20.2"
+  },
+  "devDependencies": {
+    "nodemon": "^2.0.22"
+  }
+}
