+++ scripts/.placeholder
@@ -1,1 +1,1 @@
-# This is a placeholder file to create the scripts directory 
+# Placeholder for PowerShell scripts
