--- web-ui/.gitignore
@@ -1,106 +1,0 @@
-# Dependency directories
-node_modules/
-jspm_packages/
-
-# Environment variables
-.env
-.env.local
-.env.development.local
-.env.test.local
-.env.production.local
-
-# Build output
-dist/
-build/
-coverage/
-
-# Logs
-logs
-*.log
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-lerna-debug.log*
-.pnpm-debug.log*
-
-# Cache directories
-.npm
-.eslintcache
-.stylelintcache
-.node_repl_history
-
-# Runtime data
-pids
-*.pid
-*.seed
-*.pid.lock
-
-# Debugging
-.nyc_output
-
-# IDE - VSCode
-.vscode/*
-!.vscode/settings.json
-!.vscode/tasks.json
-!.vscode/launch.json
-!.vscode/extensions.json
-.history/
-
-# IDE - JetBrains (WebStorm, IntelliJ, etc)
-.idea/
-*.iml
-*.iws
-.idea_modules/
-
-# IDE - Visual Studio
-*.suo
-*.ntvs*
-*.njsproj
-*.sln
-*.sw?
-
-# macOS specific
-.DS_Store
-.AppleDouble
-.LSOverride
-._*
-
-# Windows specific
-Thumbs.db
-Thumbs.db:encryptable
-ehthumbs.db
-ehthumbs_vista.db
-*.stackdump
-[Dd]esktop.ini
-$RECYCLE.BIN/
-*.lnk
-
-# Linux specific
-*~
-.fuse_hidden*
-.directory
-.Trash-*
-.nfs*
-
-# Temporary files
-*.tmp
-*.temp
-tmp/
-temp/
-
-# Package files
-*.tgz
-*.gz
-
-# TypeScript
-*.tsbuildinfo
-
-# Optional npm cache directory
-.npm
-
-# Optional REPL history
-.node_repl_history
-
-# Output of 'npm pack'
-*.tgz
-
