--- web-ui/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "visio-temp-file-remover",
-  "version": "1.0.0",
-  "description": "A web tool for finding and deleting corrupted Visio temporary files (~$$*.~vssx) that cause errors",
-  "main": "app.js",
-  "scripts": {
-    "start": "node app.js",
-    "dev": "nodemon app.js",
-    "test": "echo \"No tests yet\" && exit 0"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/yourusername/visio-temp-file-remover.git"
-  },
-  "keywords": [
-    "visio",
-    "file-management",
-    "temporary-files",
-    "cleanup",
-    "powershell"
-  ],
-  "author": "Your Name",
-  "license": "MIT",
-  "dependencies": {
-    "express": "^4.18.2",
-    "body-parser": "^1.20.2",
-    "chalk": "^4.1.2"
-  },
-  "devDependencies": {
-    "nodemon": "^2.0.22"
-  }
-}
