--- web-ui/scripts/.placeholder
@@ -1,1 +1,0 @@
-# Placeholder for PowerShell scripts
